requirements.txt
discord.py
py2neo